Allowed userfile access by group. Modified data provider and remote resource to filter groups in the sql query. Updated documentation.
git-svn-id: svn+ssh://huia.bic.mni.mcgill.ca/home/cbrain/cbrain_repo/trunk@343 bcc1f7e1-216c-4ce3-888e-c5a8bb225047

+++ BrainPortal/app/controllers/bourreaux_controller.rb
@@ -19,10 +19,7 @@ class BourreauxController < ApplicationController
   before_filter :admin_role_required, :except => [:index]  
    
   def index #:nodoc:
-    @bourreaux = Bourreau.all;
-    if ! check_role(:admin)
-      @bourreaux = @bourreaux.select { |p| p.can_be_accessed_by(current_user) }
-    end
+    @bourreaux = Bourreau.find_all_accessible_by_user(current_user)
   end
 
   
@@ -43,6 +40,8 @@ class BourreauxController < ApplicationController
   def edit #:nodoc:
     @user     = current_user
     @bourreau = Bourreau.find(params[:id])
+    
+    raise "Bourreau not accessible by current user." unless @bourreau.can_be_accessed_by(current_user)
 
     respond_to do |format|
       format.html { render :action => :edit }
@@ -91,7 +90,9 @@ class BourreauxController < ApplicationController
   def update #:nodoc:
     @user     = current_user
     id        = params[:id]
-    @bourreau = Bourreau.find_by_id(id)
+    @bourreau = Bourreau.find(id)
+    
+    raise "Bourreau not accessible by current user." unless @bourreau.can_be_accessed_by(current_user)
 
     fields    = params[:bourreau]
     subtype   = fields.delete(:type)
@@ -115,7 +116,9 @@ class BourreauxController < ApplicationController
   def destroy #:nodoc:
     id        = params[:id]
     @user     = current_user
-    @bourreau = Bourreau.find_by_id(id)
+    @bourreau = Bourreau.find(id)
+    
+    raise "Bourreau not accessible by current user." unless @bourreau.can_be_accessed_by(current_user)
 
     if @bourreau.destroy
       flash[:notice] = "Bourreau successfully deleted."
